requirements.txt
# IELTS Score Analyzer - Dependencies

# GUI Framework (choose one)
PyQt6==6.6.1

# Optional: AI Integration
openai==1.6.1
anthropic==0.8.1
google-genai>=1.0.0

# Utilities
python-dotenv==1.0.0

# Web version (optional)
flask==3.0.0
flask-cors==4.0.0
